chore: rename properties panel
