+++ frontend/src/components/PropertiesPanel.vue
